LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/window-onerror-runtime-error-throw-expected.txt

CONSOLE MESSAGE: bar

PASS correct number of calls to window.onerror 
